requirements.txt
Flask==0.12.2
Gunicorn==19.7.1
Pillow==4.1.1
Requests==2.14.2
Scrapy==1.4.0
Werkzeug==0.12.2
